add corr matrix functions
